+++ vms/platformvm/state/state.go
@@ -33,6 +33,7 @@ import (
 	"github.com/ava-labs/avalanchego/utils"
 	"github.com/ava-labs/avalanchego/utils/constants"
 	"github.com/ava-labs/avalanchego/utils/crypto/bls"
+	"github.com/ava-labs/avalanchego/utils/hashing"
 	"github.com/ava-labs/avalanchego/utils/set"
 	"github.com/ava-labs/avalanchego/utils/units"
 	"github.com/ava-labs/avalanchego/utils/wrappers"
@@ -1277,6 +1278,129 @@ func (s *state) loadPendingStakers() error {
 	return iter.Error()
 }
 
+// Invariant: initValidatorSets requires loadCurrentValidators to have already
+// been called.
+func (s *state) initValidatorSets() error {
+	for subnetID, validators := range s.currentStakers.validators {
+		if s.validators.Count(subnetID) != 0 {
+			// Enforce the invariant that the validator set is empty here.
+			return fmt.Errorf("%w: %s", errValidatorSetAlreadyPopulated, subnetID)
+		}
+
+		for nodeID, validator := range validators {
+			validatorStaker := validator.validator
+			if err := s.validators.AddStaker(subnetID, nodeID, validatorStaker.PublicKey, validatorStaker.TxID, validatorStaker.Weight); err != nil {
+				return err
+			}
+
+			delegatorIterator := NewTreeIterator(validator.delegators)
+			for delegatorIterator.Next() {
+				delegatorStaker := delegatorIterator.Value()
+				if err := s.validators.AddWeight(subnetID, nodeID, delegatorStaker.Weight); err != nil {
+					delegatorIterator.Release()
+					return err
+				}
+			}
+			delegatorIterator.Release()
+		}
+	}
+
+	s.metrics.SetLocalStake(s.validators.GetWeight(constants.PrimaryNetworkID, s.ctx.NodeID))
+	totalWeight, err := s.validators.TotalWeight(constants.PrimaryNetworkID)
+	if err != nil {
+		return fmt.Errorf("failed to get total weight of primary network validators: %w", err)
+	}
+	s.metrics.SetTotalStake(totalWeight)
+	return nil
+}
+
+func (s *state) write(updateValidators bool, height uint64) error {
+	currentData, weightDiffs, blsKeyDiffs, valSetDiff, err := s.processCurrentStakers()
+	if err != nil {
+		return err
+	}
+	pendingData, err := s.processPendingStakers()
+	if err != nil {
+		return err
+	}
+
+	return utils.Err(
+		s.writeMerkleState(currentData, pendingData),
+		s.writeBlocks(),
+		s.writeTxs(),
+		s.writeLocalUptimes(),
+		s.writeWeightDiffs(height, weightDiffs),
+		s.writeBlsKeyDiffs(height, blsKeyDiffs),
+		s.writeRewardUTXOs(),
+		s.updateValidatorSet(updateValidators, valSetDiff, weightDiffs),
+	)
+}
+
+func (s *state) Close() error {
+	return utils.Err(
+		s.flatValidatorWeightDiffsDB.Close(),
+		s.flatValidatorPublicKeyDiffsDB.Close(),
+		s.localUptimesDB.Close(),
+		s.indexedUTXOsDB.Close(),
+		s.txDB.Close(),
+		s.blockDB.Close(),
+		s.blockIDDB.Close(),
+		s.merkleDB.Close(),
+		s.baseMerkleDB.Close(),
+	)
+}
+
+// If [ms] isn't initialized, initializes it with [genesis].
+// Then loads [ms] from disk.
+func (s *state) sync(genesis []byte) error {
+	shouldInit, err := s.shouldInit()
+	if err != nil {
+		return fmt.Errorf(
+			"failed to check if the database is initialized: %w",
+			err,
+		)
+	}
+
+	// If the database is empty, create the platform chain anew using the
+	// provided genesis state
+	if shouldInit {
+		if err := s.init(genesis); err != nil {
+			return fmt.Errorf(
+				"failed to initialize the database: %w",
+				err,
+			)
+		}
+	}
+
+	return s.load(shouldInit)
+}
+
+// Creates a genesis from [genesisBytes] and initializes [ms] with it.
+func (s *state) init(genesisBytes []byte) error {
+	// Create the genesis block and save it as being accepted (We don't do
+	// genesisBlock.Accept() because then it'd look for genesisBlock's
+	// non-existent parent)
+	genesisID := hashing.ComputeHash256Array(genesisBytes)
+	genesisBlock, err := block.NewApricotCommitBlock(genesisID, 0 /*height*/)
+	if err != nil {
+		return err
+	}
+
+	genesisState, err := genesis.Parse(genesisBytes)
+	if err != nil {
+		return err
+	}
+	if err := s.syncGenesis(genesisBlock, genesisState); err != nil {
+		return err
+	}
+
+	if err := s.doneInit(); err != nil {
+		return err
+	}
+
+	return s.Commit()
+}
+
 func (s *state) GetCurrentDelegatorIterator(subnetID ids.ID, nodeID ids.NodeID) (StakerIterator, error) {
 	return s.currentStakers.GetDelegatorIterator(subnetID, nodeID), nil
 }
@@ -1539,42 +1663,6 @@ func (s *state) CommitBatch() (database.Batch, error) {
 
 func (*state) Checksum() ids.ID {
 	return ids.Empty
-}
-
-func (s *state) Close() error {
-	return utils.Err(
-		s.flatValidatorWeightDiffsDB.Close(),
-		s.flatValidatorPublicKeyDiffsDB.Close(),
-		s.localUptimesDB.Close(),
-		s.indexedUTXOsDB.Close(),
-		s.txDB.Close(),
-		s.blockDB.Close(),
-		s.blockIDDB.Close(),
-		s.merkleDB.Close(),
-		s.baseMerkleDB.Close(),
-	)
-}
-
-func (s *state) write(updateValidators bool, height uint64) error {
-	currentData, weightDiffs, blsKeyDiffs, valSetDiff, err := s.processCurrentStakers()
-	if err != nil {
-		return err
-	}
-	pendingData, err := s.processPendingStakers()
-	if err != nil {
-		return err
-	}
-
-	return utils.Err(
-		s.writeMerkleState(currentData, pendingData),
-		s.writeBlocks(),
-		s.writeTxs(),
-		s.writeLocalUptimes(),
-		s.writeWeightDiffs(height, weightDiffs),
-		s.writeBlsKeyDiffs(height, blsKeyDiffs),
-		s.writeRewardUTXOs(),
-		s.updateValidatorSet(updateValidators, valSetDiff, weightDiffs),
-	)
 }
 
 func (s *state) processCurrentStakers() (
